+++ RELEASE_NOTES.md
@@ -6,6 +6,10 @@ _Date TBD_
 ### Changes
 * Find any current products that have bad values for their criterions' Privacy and Security Framework
 * Update values for CQMs with typos
+* Add tables for recipients and subscriptions to different types of notifications
+
+---
+
 
 ## Version 8.5.0
 _24 April 2017_